More fixes for GetData typing
