+++ src/execution/operator/join/physical_iejoin.cpp
@@ -255,7 +255,7 @@ struct IEJoinUnion {
 		}
 	}
 
-	template <typename T>
+	template <typename T, typename VECTOR_TYPE = T>
 	static vector<T> ExtractColumn(SortedTable &table, idx_t col_idx) {
 		vector<T> result;
 		result.reserve(table.count);
@@ -274,8 +274,10 @@ struct IEJoinUnion {
 				break;
 			}
 
-			const auto data_ptr = FlatVector::GetData<T>(payload.data[col_idx]);
-			result.insert(result.end(), data_ptr, data_ptr + count);
+			const auto data_ptr = FlatVector::GetData<VECTOR_TYPE>(payload.data[col_idx]);
+			for (idx_t i = 0; i < count; i++) {
+				result.push_back(UnsafeNumericCast<T>(data_ptr[i]));
+			}
 		}
 
 		return result;
@@ -513,7 +515,7 @@ IEJoinUnion::IEJoinUnion(ClientContext &context, const PhysicalIEJoin &op, Sorte
 	// We don't actually need the L2 column, just its sort key, which is in the sort blocks
 
 	// 6. compute the permutation array P of L2 w.r.t. L1
-	p = ExtractColumn<idx_t>(*l2, types.size() - 1);
+	p = ExtractColumn<idx_t, int64_t>(*l2, types.size() - 1);
 
 	// 7. initialize bit-array B (|B| = n), and set all bits to 0
 	n = l2->count.load();
